Split CLI commands into separate .py files
