+++ src/beneuro_data/cli/config_commands.py
@@ -1,0 +1,77 @@
+from pathlib import Path
+
+import typer
+from rich import print
+
+from beneuro_data.config import _get_env_path, _load_config
+
+
+def show_config():
+    """
+    Show the contents of the config file.
+    """
+    config = _load_config()
+    print(config.json(indent=4))
+
+
+def _check_root(root_path: Path):
+    assert root_path.exists(), f"{root_path} does not exist."
+    assert root_path.is_dir(), f"{root_path} is not a directory."
+
+    files_in_root = [f.stem for f in root_path.iterdir()]
+
+    assert "raw" in files_in_root, f"No raw folder in {root_path}"
+    assert "processed" in files_in_root, f"No processed folder in {root_path}"
+
+
+def check_config():
+    """
+    Check that the local and remote root folders have the expected raw and processed folders.
+    """
+    config = _load_config()
+
+    print(
+        "Checking that local and remote root folders have the expected raw and processed folders..."
+    )
+
+    _check_root(config.LOCAL_PATH)
+    _check_root(config.REMOTE_PATH)
+
+    print("[green]Config looks good.")
+
+
+def init():
+    """
+    Create a .env file to store the paths to the local and remote data storage.
+    """
+
+    # check if the file exists
+    env_path = _get_env_path()
+
+    if env_path.exists():
+        print("\n[yellow]Config file already exists.\n")
+
+        check_config()
+
+    else:
+        print("\nConfig file doesn't exist. Let's create one.")
+
+        local_path = Path(
+            typer.prompt("Enter the absolute path to the root of the local data storage")
+        )
+        _check_root(local_path)
+        remote_path = Path(
+            typer.prompt("Enter the absolute path to the root of remote data storage")
+        )
+        _check_root(remote_path)
+
+        with open(env_path, "w") as f:
+            f.write(f"LOCAL_PATH = {local_path}\n")
+            f.write(f"REMOTE_PATH = {remote_path}\n")
+
+        # make sure that it works
+        config = _load_config()
+        _check_root(config.LOCAL_PATH)
+        _check_root(config.REMOTE_PATH)
+
+        print("[green]Config file created successfully.")
