+++ src/beneuro_data/cli/rename_videos_command.py
@@ -1,0 +1,48 @@
+from pathlib import Path
+from typing_extensions import Annotated
+
+import typer
+
+from beneuro_data.video_renaming import rename_raw_videos_of_session
+
+
+def rename_videos(
+    session_path: Annotated[
+        Path, typer.Argument(help="Path to session directory. Can be relative or absolute")
+    ],
+    subject_name: Annotated[
+        str,
+        typer.Argument(
+            help="Name of the subject the session belongs to. (Used for confirmation.)"
+        ),
+    ],
+    processing_level: Annotated[
+        str, typer.Argument(help="Processing level of the session. raw or processed.")
+    ] = "raw",
+    verbose: Annotated[
+        bool,
+        typer.Option(help="Print the list of files that were renamed."),
+    ] = False,
+):
+    """
+    Rename the raw videos saved by Jarvis during a session to the convention we use.
+
+    Example usage:
+
+        `bnd rename-videos . M017 --verbose`
+
+        `bnd rename-videos /absolute/path/to/session M017 --verbose`
+    """
+    if processing_level != "raw":
+        raise NotImplementedError("Sorry, only raw data is supported for now.")
+
+    if not session_path.absolute().is_dir():
+        raise ValueError("Session path must be a directory.")
+    if not session_path.absolute().exists():
+        raise ValueError("Session path does not exist.")
+
+    rename_raw_videos_of_session(
+        session_path.absolute(),
+        subject_name,
+        verbose,
+    )
